requirements.txt
async-timeout==4.0.3
greenlet==3.0.1
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
